Draft for IP camera handling: new module onvif_cam_management + changes in local_ffmpeg_hls Removed some redundant configs from modules configs files (using global configs instead) Improved installation process a bit, more to do there. File some php warnings about everywhere. Moved all model functions from web_index.php to lib_model.php

+++ modules/local_ffmpeg_hls/lib_capture.php
@@ -24,8 +24,10 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-require 'config.inc';
+require 'etc/config.inc';
+require_once __DIR__ . '/../../global_config.inc';
 require_once $basedir . '/lib_various.php';
+require_once 'create_bash_configs.php';
 include_once $basedir . '/lib_error.php';
 
 /*
@@ -41,22 +43,33 @@ include_once $basedir . '/lib_error.php';
  * This function should be called before the use of the camera.
  * This function should launch a background task to save time and keep syncro
  * between cam and slides (if both are available)
- * @param int $pid the process id of the background task
+ * @param int $pid the process id of the background task. This is updated to process pid if function is successfull
  * @param associative_array $meta_assoc Metadata related to the record (used in cli_monitoring.php)
  * @return boolean true if everything went well; false otherwise
  */
 function capture_ffmpeg_init(&$pid, $meta_assoc) {
-    global $ffmpeg_script_init; 
+    global $ffmpeg_script_init;
     global $ffmpeg_recorder_logs;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
     global $ezcast_submit_url;
     global $ffmpeg_streaming_info;
     global $ffmpeg_streaming_quality;
     global $ffmpeg_input_source;
     global $php_cli_cmd;
     global $ffmpeg_cli_streaming;
+    global $bash_env;
+    global $ffmpeg_basedir;
 
-    unlink($ffmpeg_streaming_info);
+    //prepare bash variables
+    $success = create_bash_configs($bash_env, $ffmpeg_basedir . "etc/localdefs");
+    if (!$success) {
+        file_put_contents($ffmpeg_recorder_logs, "capture_ffmpeg_init: ERROR: Unable to create bash variables file\n", FILE_APPEND);
+        return false;
+    }
+
+    if (file_exists($ffmpeg_streaming_info))
+        unlink($ffmpeg_streaming_info);
+
     $asset = $meta_assoc['record_date'] . '_' . $meta_assoc['course_name'];
 
     $tmp_dir = capture_ffmpeg_tmpdir_get($asset);
@@ -64,53 +77,57 @@ function capture_ffmpeg_init(&$pid, $meta_assoc) {
     // saves recording metadata as xml file 
     assoc_array2xml_file($meta_assoc, "$tmp_dir/_metadata.xml");
 
-
     // status of the current recording
     $status = capture_ffmpeg_status_get();
-    if ($status == '') { // no status yet
-        $streaming_info = capture_ffmpeg_info_get('streaming', $asset);
-        if ($streaming_info !== false){
-            // defines that the streaming is enabled
-            // It must be done before calling $ffmpeg_script_init (for preparing low and high HLS streams)
-            file_put_contents($ffmpeg_streaming_info, var_export($streaming_info, true));
-        }
-        // script_init initializes FFMPEG and launches the recording
-        // in background to save time (pid is returned to be handled by web_index.php)
-        system("sudo -u $ffmpeg_username $ffmpeg_script_init $asset $ffmpeg_input_source 1 >> $ffmpeg_recorder_logs 2>&1 & echo $! > $tmp_dir/pid");
-        $pid = file_get_contents("$tmp_dir/pid");
-        // error occured while launching FMLE
-        if (capture_ffmpeg_status_get() == 'launch_failure') {
-            error_last_message("can't open because FFMPEG failed to launch");
-            return false;
-        }
-        // the recording is now 'open'
-        capture_ffmpeg_status_set('open');
 
-        // init the streaming
-        if ($streaming_info !== false) {
-            // streaming is enabled, we send a request to EZmanager to 
-            // init the streamed asset
-            $post_array = $streaming_info;
-            $post_array['action'] = 'streaming_init';
-            $result = server_request_send($ezcast_submit_url, $post_array);
-
-            if (strpos($result, 'Curl error') !== false) {
-                // an error occured with CURL
-                $meta_assoc['streaming'] = 'false';
-                unlink($ffmpeg_streaming_info);
-            }
-            $result = unserialize($result);
-
-            // executes the command for sending TS segments to EZmanager in background
-            // for low and high qualities
-            if (strpos($ffmpeg_streaming_quality, 'high') !== false)
-                exec("$php_cli_cmd $ffmpeg_cli_streaming " . $meta_assoc['course_name'] . " " . $meta_assoc['record_date'] . " high > /dev/null &", $output, $errno);
-            if (strpos($ffmpeg_streaming_quality, 'low') !== false)
-                exec("$php_cli_cmd $ffmpeg_cli_streaming " . $meta_assoc['course_name'] . " " . $meta_assoc['record_date'] . " low > /dev/null &", $output, $errno);
-        }
-    } else {
+    if ($status != '') { // has a status
         error_last_message("capture_init: can't open because current status: $status");
         return false;
+    }
+
+    // no status yet
+    $streaming_info = capture_ffmpeg_info_get('streaming', $asset);
+    if ($streaming_info !== false) {
+        // defines that the streaming is enabled
+        // It must be done before calling $ffmpeg_script_init (for preparing low and high HLS streams)
+        file_put_contents($ffmpeg_streaming_info, var_export($streaming_info, true));
+    }
+    // script_init initializes FFMPEG and launches the recording
+    // in background to save time (pid is returned to be handled by web_index.php)
+    system("sudo -u $ezrecorder_username $ffmpeg_script_init $asset $ffmpeg_input_source 1 >> $ffmpeg_recorder_logs 2>&1 & echo $! > $tmp_dir/pid");
+    $pid = file_get_contents("$tmp_dir/pid");
+    // error occured while launching FMLE
+    if (capture_ffmpeg_status_get() == 'launch_failure') {
+        error_last_message("can't open because FFMPEG failed to launch");
+        return false;
+    }
+    // the recording is now 'open'
+    capture_ffmpeg_status_set('open');
+
+    // init the streaming
+    if ($streaming_info !== false) {
+        // streaming is enabled, we send a request to EZmanager to 
+        // init the streamed asset
+        $post_array = $streaming_info;
+        $post_array['action'] = 'streaming_init';
+        $result = server_request_send($ezcast_submit_url, $post_array);
+
+        if (strpos($result, 'Curl error') !== false) {
+            // an error occured with CURL
+            $meta_assoc['streaming'] = 'false';
+            unlink($ffmpeg_streaming_info);
+        }
+        //not used $result = unserialize($result);
+        // executes the command for sending TS segments to EZmanager in background
+        // for low and high qualities
+        if (strpos($ffmpeg_streaming_quality, 'high') !== false) {
+            exec("$php_cli_cmd $ffmpeg_cli_streaming " . $meta_assoc['course_name']
+                    . " " . $meta_assoc['record_date'] . " high > /dev/null &", $output, $errno);
+        }
+        if (strpos($ffmpeg_streaming_quality, 'low') !== false) {
+            exec("$php_cli_cmd $ffmpeg_cli_streaming " . $meta_assoc['course_name']
+                    . " " . $meta_assoc['record_date'] . " low > /dev/null &", $output, $errno);
+        }
     }
 
     return true;
@@ -123,11 +140,11 @@ function capture_ffmpeg_init(&$pid, $meta_assoc) {
 function capture_ffmpeg_start($asset) {
     global $ffmpeg_script_start;
     global $ffmpeg_recorder_logs;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
     global $ffmpeg_input_source;
 
     // adds time in the cutlist 
-    system("sudo -u $ffmpeg_username $ffmpeg_script_start $asset $ffmpeg_input_source >> $ffmpeg_recorder_logs 2>&1 &");
+    system("sudo -u $ezrecorder_username $ffmpeg_script_start $asset $ffmpeg_input_source >> $ffmpeg_recorder_logs 2>&1 &");
 
     //update recording status
     $status = capture_ffmpeg_status_get();
@@ -149,13 +166,13 @@ function capture_ffmpeg_start($asset) {
 function capture_ffmpeg_pause($asset) {
     global $ffmpeg_script_cutlist;
     global $ffmpeg_recorder_logs;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
 
     // get status of the current recording
     $status = capture_ffmpeg_status_get();
     if ($status == 'recording') {
         // qtbpause pauses the recording in QTB
-        system("sudo -u $ffmpeg_username $ffmpeg_script_cutlist $asset pause >> $ffmpeg_recorder_logs 2>&1 &");
+        system("sudo -u $ezrecorder_username $ffmpeg_script_cutlist $asset pause >> $ffmpeg_recorder_logs 2>&1 &");
         capture_ffmpeg_status_set('paused');
     } else {
         error_last_message("capture_pause: can't pause recording because current status: $status");
@@ -172,13 +189,13 @@ function capture_ffmpeg_pause($asset) {
 function capture_ffmpeg_resume($asset) {
     global $ffmpeg_script_cutlist;
     global $ffmpeg_recorder_logs;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
 
     // get status of the current recording
     $status = capture_ffmpeg_status_get();
     if ($status == 'paused' || $status == 'stopped') {
         // qtbresume resumes the current recording
-        system("sudo -u $ffmpeg_username $ffmpeg_script_cutlist $asset resume >> $ffmpeg_recorder_logs 2>&1 &");
+        system("sudo -u $ezrecorder_username $ffmpeg_script_cutlist $asset resume >> $ffmpeg_recorder_logs 2>&1 &");
         // sets the new status of the current recording
         capture_ffmpeg_status_set('recording');
     } else {
@@ -196,7 +213,7 @@ function capture_ffmpeg_resume($asset) {
 function capture_ffmpeg_stop(&$pid, $asset) {
     global $ffmpeg_script_cutlist;
     global $ffmpeg_recorder_logs;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
 
     $tmp_dir = capture_ffmpeg_tmpdir_get($asset);
 
@@ -204,7 +221,7 @@ function capture_ffmpeg_stop(&$pid, $asset) {
     $status = capture_ffmpeg_status_get();
     if ($status == 'recording' || $status == "paused") {
         // pauses the current recording (while user chooses the way to publish the record)
-        system("sudo -u $ffmpeg_username $ffmpeg_script_cutlist $asset stop >> $ffmpeg_recorder_logs 2>&1 & echo $! > $tmp_dir/pid");
+        system("sudo -u $ezrecorder_username $ffmpeg_script_cutlist $asset stop >> $ffmpeg_recorder_logs 2>&1 & echo $! > $tmp_dir/pid");
         $pid = file_get_contents("$tmp_dir/pid");
 
         // set the new status for the current recording
@@ -225,13 +242,13 @@ function capture_ffmpeg_cancel($asset) {
 
     global $ffmpeg_script_cancel;
     global $ffmpeg_recorder_logs;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
 
     // get status of the current recording
     $status = capture_ffmpeg_status_get();
     if ($status == 'recording' || $status == 'stopped' || $status == 'paused' || $status == 'open' || $status == '') {
         // qtbcancel cancels the current recording, saves it in archive dir and stops the monitoring
-        $cmd = 'sudo -u ' . $ffmpeg_username . ' ' . $ffmpeg_script_cancel . ' ' . $asset . ' >> ' . $ffmpeg_recorder_logs . ' 2>&1 &';
+        $cmd = 'sudo -u ' . $ezrecorder_username . ' ' . $ffmpeg_script_cancel . ' ' . $asset . ' >> ' . $ffmpeg_recorder_logs . ' 2>&1 &';
         log_append('recording', "launching command: $cmd");
         $fpart = exec($cmd, $outputarray, $errorcode);
         $post_array = capture_ffmpeg_info_get('streaming', $asset);
@@ -263,7 +280,7 @@ function capture_ffmpeg_process($meta_assoc, &$pid) {
     global $ffmpeg_recorder_logs;
     global $ffmpeg_processing_tool;
     global $ffmpeg_processing_tools;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
 
     $asset = $meta_assoc['record_date'] . '_' . $meta_assoc['course_name'];
     $tmp_dir = capture_ffmpeg_tmpdir_get($asset);
@@ -278,7 +295,7 @@ function capture_ffmpeg_process($meta_assoc, &$pid) {
     if ($status != 'recording' && $status != 'open') {
         // saves recording in processing dir and processes it
         // launched in background to save time 
-        $cmd = 'sudo -u ' . $ffmpeg_username . ' ' . $ffmpeg_script_stop . ' ' . $meta_assoc['course_name'] . ' ' . $meta_assoc['record_date'] . ' ' . $ffmpeg_processing_tool . ' >> ' . $ffmpeg_recorder_logs . ' 2>&1  & echo $! > ' . $tmp_dir . '/pid';
+        $cmd = 'sudo -u ' . $ezrecorder_username . ' ' . $ffmpeg_script_stop . ' ' . $meta_assoc['course_name'] . ' ' . $meta_assoc['record_date'] . ' ' . $ffmpeg_processing_tool . ' >> ' . $ffmpeg_recorder_logs . ' 2>&1  & echo $! > ' . $tmp_dir . '/pid';
         log_append('recording', "launching command: $cmd");
         // returns the process id of the background task
         system($cmd);
@@ -325,7 +342,7 @@ function capture_ffmpeg_process($meta_assoc, &$pid) {
 function capture_ffmpeg_finalize($asset) {
     global $ffmpeg_script_finalize;
     global $ffmpeg_recorder_logs;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
 
     $tmp_dir = capture_ffmpeg_tmpdir_get($asset);
 
@@ -333,7 +350,7 @@ function capture_ffmpeg_finalize($asset) {
     $meta_assoc = xml_file2assoc_array("$tmp_dir/_metadata.xml");
 
     // launches finalization bash script
-    $cmd = 'sudo -u ' . $ffmpeg_username . ' ' . $ffmpeg_script_finalize . ' ' . $meta_assoc['course_name'] . " " . $meta_assoc['record_date'] . ' >> ' . $ffmpeg_recorder_logs . ' 2>&1  & echo $!';
+    $cmd = 'sudo -u ' . $ezrecorder_username . ' ' . $ffmpeg_script_finalize . ' ' . $meta_assoc['course_name'] . " " . $meta_assoc['record_date'] . ' >> ' . $ffmpeg_recorder_logs . ' 2>&1  & echo $!';
     log_append("finalizing: execute cmd '$cmd'");
     $res = exec($cmd, $output, $errorcode);
 }
@@ -341,17 +358,17 @@ function capture_ffmpeg_finalize($asset) {
 /**
  * @implements
  * Returns an associative array containing information required for given action
- * @global type $ffmpeg_ip
+ * @global type $ezrecorder_ip
  * @global type $ffmpeg_download_protocol
- * @global type $ffmpeg_username
+ * @global type $ezrecorder_username
  * @return type
  */
 function capture_ffmpeg_info_get($action, $asset = '') {
-    global $ffmpeg_ip;
+    global $ezrecorder_ip;
     global $ffmpeg_download_protocol;
     global $ffmpeg_streaming_protocol;
     global $ffmpeg_streaming_quality;
-    global $ffmpeg_username;
+    global $ezrecorder_username;
     global $ffmpeg_upload_dir;
     global $classroom;
     global $cam_module;
@@ -362,16 +379,17 @@ function capture_ffmpeg_info_get($action, $asset = '') {
             $meta_assoc = xml_file2assoc_array("$tmp_dir/_metadata.xml");
 
             // rsync requires ssh protocol is set (key sharing) on the remote server
-            $download_info_array = array("ip" => $ffmpeg_ip,
+            $download_info_array = array("ip" => $ezrecorder_ip,
                 "protocol" => $ffmpeg_download_protocol,
-                "username" => $ffmpeg_username,
+                "username" => $ezrecorder_username,
                 "filename" => $ffmpeg_upload_dir . $meta_assoc['record_date'] . "_" . $meta_assoc['course_name'] . "/cam.mov");
             return $download_info_array;
             break;
         case 'streaming':
             include_once 'info.php';
             $tmp_dir = capture_ffmpeg_tmpdir_get($asset);
-            if ($ffmpeg_streaming_quality == 'none') return false;
+            if ($ffmpeg_streaming_quality == 'none')
+                return false;
             $meta_assoc = xml_file2assoc_array("$tmp_dir/_metadata.xml");
             $module_type = (($cam_module == $module_name) ? 'cam' : 'slide');
             // streaming is disabled if it has not been enabled by user
@@ -379,12 +397,12 @@ function capture_ffmpeg_info_get($action, $asset = '') {
             if ($meta_assoc['streaming'] === 'false' || ($meta_assoc['record_type'] !== 'camslide' && $meta_assoc['record_type'] != $module_type))
                 return false;
             $streaming_info_array = array(
-                "ip" => $ffmpeg_ip,
+                "ip" => $ezrecorder_ip,
                 "protocol" => $ffmpeg_streaming_protocol,
                 "album" => $meta_assoc['course_name'],
                 "asset" => $meta_assoc['record_date'],
                 "record_type" => $meta_assoc['record_type'],
-                "module_type" => $module_type, 
+                "module_type" => $module_type,
                 "module_quality" => $ffmpeg_streaming_quality,
                 "classroom" => $classroom,
                 "netid" => $meta_assoc['netid'],
@@ -406,8 +424,8 @@ function capture_ffmpeg_thumbnail() {
 
     // Camera screenshot
     $diff = time() - filemtime($ffmpeg_capture_file);
-    if (!file_exists($ffmpeg_capture_file) || (time() - filemtime($ffmpeg_capture_file) > 3)) {
-        //if no image or image is old get a new screecapture
+    if (!file_exists($ffmpeg_capture_file) || ($diff > 3)) {
+        //if no image or image is old get a new screencapture
         if ((time() - filemtime("$ffmpeg_basedir/var/pic_new.jpg") > 3)) {
             //print "could not take a screencapture";
             copy("./nopic.jpg", $ffmpeg_capture_file);
@@ -417,8 +435,7 @@ function capture_ffmpeg_thumbnail() {
             if ($status == 'recording') {
                 $status = capture_ffmpeg_recstatus_get();
             }
-            image_resize("$ffmpeg_basedir/var/pic_new.jpg", "$ffmpeg_basedir/var/pic_new_www.jpg", 235, 157, $status, false);
-            rename("$ffmpeg_basedir/var/pic_new_www.jpg", $ffmpeg_capture_file);
+            rename("$ffmpeg_basedir/var/pic_new.jpg", $ffmpeg_capture_file);
         }
     }
     return file_get_contents($ffmpeg_capture_file);
@@ -457,19 +474,18 @@ function capture_ffmpeg_status_set($status) {
 function capture_ffmpeg_features_get() {
     global $ffmpeg_features;
     global $ffmpeg_streaming_quality;
-    
-    if ($ffmpeg_streaming_quality == 'none'){
-        if($index = array_search('streaming', $ffmpeg_features) !== false){
+
+    if ($ffmpeg_streaming_quality == 'none') {
+        if ($index = array_search('streaming', $ffmpeg_features) !== false) {
             unset($ffmpeg_features[$index]);
         }
     }
-    
+
     return $ffmpeg_features;
 }
 
 /**
  * returns the real status of the current recording
- * @global type $ffmpeg_recorddir
  * @global type $ffmpeg_movie_name 
  * @return string
  */
@@ -498,7 +514,7 @@ function capture_ffmpeg_tmpdir_get($asset) {
     static $tmp_dir;
 
     $tmp_dir = $ffmpeg_basedir . '/var/' . $asset;
-    if (!dir($tmp_dir))
+    if (!file_exists($tmp_dir))
         mkdir($tmp_dir, 0777, true);
 
     return $tmp_dir;
